+++ docker/db/sql/3_create_authors_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS `authors` (
-  `id` int(5) NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `name` varchar(50),
   PRIMARY KEY (`id`,`name`),
   UNIQUE KEY `name` (`name`)
